# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Detail": {
        "description": "FastAPI's error body, written down so it can be named in `responses=`.\n\nEvery non-validation error this API returns is one of these. Declaring it\nis what turns \"401\" in the docs from a bare number into a shape a client\ncan code against.",
        "example": {
          "detail": "Could not validate credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Detail",
        "type": "object"
      },
...
      "MeOut": {
        "description": "A person, as they see themselves. Same shape plus the email, because on\nthis one endpoint the caller is the only person it belongs to.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at",
          "email"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "MeUpdate": {
        "description": "What you may change about yourself. One field, on purpose.\n\nNot the email: it is the credential half of an account, and changing one is\na flow with a confirmation link in it rather than a text box. Not the\npassword either, for the same reason. The username is the public half —\nwhat everybody else sees — and the one that is nobody's business but yours.",
        "properties": {
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "MeUpdate",
        "type": "object"
      },
...
      "Token": {
        "description": "What `/login` and `/auth/refresh` hand back.\n\nNote what *isn't* here: the refresh token. It goes out in an httpOnly\ncookie, where no script on the page can read it, which is the whole reason\nthere are two tokens rather than one long-lived one.",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjo3fQ.sig",
          "csrf_token": "0lNrK8s2Qm6Xv3dYpHtW1cJfB9aZuE4g",
          "expires_in": 900,
          "token_type": "bearer"
        },
        "properties": {
          "access_token": {
            "description": "Send as `Authorization: Bearer …`. Hold it in memory, not in storage — it expires in minutes and `/auth/refresh` replaces it.",
            "title": "Access Token",
            "type": "string"
          },
          "csrf_token": {
            "description": "Echo this back as an `X-CSRF-Token` header on `/auth/refresh` and `/auth/logout`. It changes on every rotation.",
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Seconds until the access token expires. Refresh a little before this rather than waiting for a 401.",
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "csrf_token"
        ],
        "title": "Token",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/AndrewTechTips/social-feed-application"
    },
    "description": "The backend behind **Commons** — a small public feed.\n\nPosts, comments, upvotes and accounts, with a two-token auth flow. Reading is public; writing needs an account.\n\nThe published demo at [andrewtechtips.github.io/social-feed-application](https://andrewtechtips.github.io/social-feed-application/) reimplements this same contract in the browser, because a static host has nowhere to run FastAPI. This document is the contract both of them answer to.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/AndrewTechTips/social-feed-application/blob/main/LICENSE"
    },
    "title": "Social Feed API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate the session.\n\nEvery call invalidates the token that was presented and issues the next\none, so a copied cookie is detectable: the copy eventually presents a\nsecret that has already been rotated away, and the whole session is ended\nrather than guessing which of the two holders is the thief.\n\nSend the `csrf_token` from the last `/login` or `/auth/refresh` response in\nan `X-CSRF-Token` header. Without it this answers 401 even with a perfectly\ngood cookie.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No usable refresh cookie, or no matching `X-CSRF-Token`. One answer for every way of being wrong — saying which would tell a caller whether they hold half of a real session. The cookie is cleared, because whatever is in the browser is no longer worth sending."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "Too many attempts (30/minute)."
          }
        },
        "summary": "Trade the refresh cookie for a new access token",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "delete": {
        "description": "Delete your account and everything attached to it.\n\nEverything means everything: posts, the comments on them, the comments you\nleft elsewhere, your votes, and every session you have open. None of that\nis spelled out here — each one is an `ON DELETE CASCADE` declared on the\ncolumn that points at you, so it holds whoever issues the DELETE and cannot\ndrift from what this function happens to remember to clean up.\n\nThe refresh cookie is cleared on the way out. It names a session row that\nno longer exists, so the browser would recover on its own the next time it\ntried; leaving somebody holding a credential for an account that is gone is\nnevertheless not a tidy way to say goodbye.",
        "operationId": "delete_me_api_v1_users_me_delete",
        "responses": {
          "204": {
            "description": "Gone, along with everything you wrote."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete your account",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Who the caller is.\n\nLogging in takes an email and a password and hands back a token, which\ntells the client nothing about the person it just signed in. That was\nsurvivable while posts carried their author's email and the client could\nmatch on it; now that the public identity is a username and the email never\nleaves this endpoint, something has to close the gap. Without it a browser\nwith no local history can sign in successfully and still not know which\nposts are its own.",
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-08-21T09:14:02.511Z",
                  "email": "maren.holt@example.com",
                  "id": 7,
                  "username": "marenholt"
                },
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Who the caller is",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Change the name everybody else sees.\n\nOnly the username. The email and the password are the credential half of an\naccount and changing either is a flow with a confirmation in it, not a text\nbox on a settings screen.\n\nA name that is taken is a 409 with a sentence, the same answer registering\ngives — and for the same reason it is caught from the database rather than\nchecked first: two people asking for the same free name in the same moment\nboth pass a lookup and one of them still has to lose.\n\nNothing else moves. Posts, comments and votes are joined by id, and the id\nis the half of an identity that was always meant to be the durable one.",
        "operationId": "update_me_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-08-21T09:14:02.511Z",
                  "email": "maren.holt@example.com",
                  "id": 7,
                  "username": "marenholt"
                },
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "That would collide with something already there."
          },
          "422": {
            "description": "The body or the query string didn't validate. FastAPI's own shape: `detail` is a list of `{loc, msg, type}`, one per field."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Change your username",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A local development server",
      "url": "http://localhost:8000"
    }
  ],
...
}